feat: pull up useMemo
