+++ src/components/controls/TextInput/TextInput.tsx
@@ -114,12 +114,31 @@ export const TextInput = React.forwardRef<HTMLSpanElement, TextInputProps>(funct
     const labelSize = useElementSize(isLabelVisible ? labelRef : null, size);
     const leftContentSize = useElementSize(isLeftContentVisible ? leftContentRef : null, size);
 
+    const describedBy = React.useMemo(() => {
+        const result = [];
+        if (originalControlProps?.['aria-describedby']) {
+            result.push(originalControlProps['aria-describedby']);
+        }
+        if (id) {
+            if (note) {
+                result.push(getControlNoteId(id));
+            }
+
+            if (error) {
+                result.push(getControlErrorTextId(id));
+            }
+        }
+        return result.join(' ');
+    }, [originalControlProps, id, note, error]);
+
     const controlProps: TextInputProps['controlProps'] = {
         ...originalControlProps,
         style: {
             ...originalControlProps?.style,
             ...(isLabelVisible && labelSize.width ? {paddingLeft: `${labelSize.width}px`} : {}),
         },
+        'aria-invalid': originalControlProps?.['aria-invalid'] ?? Boolean(error),
+        'aria-describedby': describedBy,
     };
     const commonProps = {
         id,
